Move RiskSection into components
